# swagger.yaml
swagger: '2.0'
info:
  title: Document Search API
  description: |
    A multi-tenant document search API built with OpenSearch and Redis caching.
    
    ## Features
    - Full-text search with BM25 ranking
    - Multi-tenant isolation
    - Redis caching for improved performance
    - Rate limiting per tenant
    
    ## Authentication
    All endpoints (except `/health`) require a tenant identifier in the `X-Tenant-Id` header.
    The tenant ID must be alphanumeric with hyphens/underscores only.
    
    ## Rate Limiting
    Rate limiting is enforced per tenant:
    - **100 requests per 60 seconds** per tenant
    
    Rate limit information is included in response headers:
    - `X-RateLimit-Limit`: Maximum requests allowed
    - `X-RateLimit-Remaining`: Remaining requests in current window
    - `X-RateLimit-Reset`: Timestamp when the rate limit resets
    - `Retry-After`: Seconds to wait before retrying (when rate limited)
  version: 1.0.0
  contact:
    name: API Support
  license:
    name: MIT

host: localhost:3000
basePath: /
schemes:
  - http
  - https

tags:
  - name: Health
    description: Health check endpoints
  - name: Documents
    description: Document management operations
  - name: Search
    description: Document search operations

securityDefinitions:
  TenantAuth:
    type: apiKey
    name: X-Tenant-Id
    in: header
    description: |
      Tenant identifier (alphanumeric with hyphens/underscores only).
      
      **Example values:**
      - `tenant-123`
      - `tenant_abc`
      - `my-tenant-001`
      
      Click the "Authorize" button above to set your tenant ID for all requests.

paths:
  /health:
    get:
      tags:
        - Health
      summary: Health check
      description: Check the health status of the service and its dependencies
      operationId: getHealth
      produces:
        - application/json
      responses:
        '200':
          description: Service is healthy
          schema:
            $ref: '#/definitions/HealthStatus'
        '503':
          description: Service is unavailable
          schema:
            $ref: '#/definitions/HealthStatus'

  /documents:
    post:
      tags:
        - Documents
      summary: Create document
      description: Index a new document for the authenticated tenant
      operationId: createDocument
      security:
        - TenantAuth: []
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/CreateDocumentRequest'
      responses:
        '201':
          description: Document successfully indexed
          schema:
            $ref: '#/definitions/CreateDocumentResponse'
        '400':
          description: Invalid request body or missing tenant ID
          schema:
            $ref: '#/definitions/Error'
        '429':
          description: Rate limit exceeded
          schema:
            $ref: '#/definitions/Error'
          headers:
            Retry-After:
              type: integer
              description: Seconds to wait before retrying

  /documents/{id}:
    get:
      tags:
        - Documents
      summary: Get document
      description: Retrieve a specific document by ID for the authenticated tenant
      operationId: getDocument
      security:
        - TenantAuth: []
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          required: true
          type: string
          description: Document ID
      responses:
        '200':
          description: Document found
          schema:
            $ref: '#/definitions/Document'
        '400':
          description: Missing tenant ID
          schema:
            $ref: '#/definitions/Error'
        '404':
          description: Document not found or belongs to different tenant
          schema:
            $ref: '#/definitions/Error'
        '429':
          description: Rate limit exceeded
          schema:
            $ref: '#/definitions/Error'
          headers:
            Retry-After:
              type: integer
              description: Seconds to wait before retrying

    delete:
      tags:
        - Documents
      summary: Delete document
      description: Delete a document by ID for the authenticated tenant
      operationId: deleteDocument
      security:
        - TenantAuth: []
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          required: true
          type: string
          description: Document ID
      responses:
        '200':
          description: Document deleted (or already deleted)
          schema:
            $ref: '#/definitions/DeleteDocumentResponse'
        '400':
          description: Missing tenant ID
          schema:
            $ref: '#/definitions/Error'
        '429':
          description: Rate limit exceeded
          schema:
            $ref: '#/definitions/Error'
          headers:
            Retry-After:
              type: integer
              description: Seconds to wait before retrying

  /search:
    get:
      tags:
        - Search
      summary: Search documents
      description: |
        Search documents for the authenticated tenant.
        
        Supports full-text search with BM25 ranking algorithm.
        Searches across title (3x boost), content, and tags (2x boost).
        
        Results are cached for 120 seconds.
      operationId: searchDocuments
      security:
        - TenantAuth: []
      produces:
        - application/json
      parameters:
        - name: q
          in: query
          type: string
          description: Search query string
        - name: offset
          in: query
          type: integer
          minimum: 0
          default: 0
          description: "Pagination offset (default: 0)"
        - name: limit
          in: query
          type: integer
          minimum: 1
          maximum: 50
          default: 10
          description: "Number of results per page (default: 10, max: 50)"
        - name: tag
          in: query
          type: string
          description: Filter by tag (exact match)
        - name: author
          in: query
          type: string
          description: Filter by author (exact match)
        - name: from
          in: query
          type: string
          format: date-time
          description: Filter by creation date from (ISO 8601 format)
        - name: to
          in: query
          type: string
          format: date-time
          description: Filter by creation date to (ISO 8601 format)
      responses:
        '200':
          description: Search completed successfully
          schema:
            $ref: '#/definitions/SearchResponse'
        '400':
          description: Invalid query parameters or missing tenant ID
          schema:
            $ref: '#/definitions/Error'
        '429':
          description: Rate limit exceeded
          schema:
            $ref: '#/definitions/Error'
          headers:
            Retry-After:
              type: integer
              description: Seconds to wait before retrying

definitions:
  HealthStatus:
    type: object
    required:
      - status
      - dependencies
    properties:
      status:
        type: string
        enum: [UP, DOWN, DEGRADED]
        description: Overall service status
      dependencies:
        type: object
        required:
          - elasticsearch
          - redis
        properties:
          elasticsearch:
            type: string
            enum: [UP, DOWN]
            description: OpenSearch/Elasticsearch status
          redis:
            type: string
            enum: [UP, DOWN]
            description: Redis cache status

  CreateDocumentRequest:
    type: object
    required:
      - title
      - content
    properties:
      id:
        type: string
        description: Optional document ID. If not provided, one will be generated.
      title:
        type: string
        minLength: 1
        maxLength: 500
        description: Document title
      content:
        type: string
        minLength: 1
        maxLength: 100000
        description: Full document content
      tags:
        type: array
        items:
          type: string
        description: Array of tags for categorization
      metadata:
        type: object
        properties:
          author:
            type: string
            description: Document author (e.g., email address)
          type:
            type: string
            description: Document type (e.g., technical_doc, incident_report)
          department:
            type: string
            description: Department or team name
        additionalProperties: true
        description: Additional metadata fields

  CreateDocumentResponse:
    type: object
    required:
      - id
      - tenantId
      - status
    properties:
      id:
        type: string
        description: Document ID
      tenantId:
        type: string
        description: Tenant ID
      status:
        type: string
        enum: [indexed]
        description: Indexing status

  Document:
    type: object
    required:
      - id
      - tenantId
      - title
      - content
    properties:
      id:
        type: string
        description: Document ID
      tenantId:
        type: string
        description: Tenant ID
      title:
        type: string
        description: Document title
      content:
        type: string
        description: Full document content
      tags:
        type: array
        items:
          type: string
        description: Array of tags
      metadata:
        type: object
        properties:
          author:
            type: string
          type:
            type: string
          department:
            type: string
        additionalProperties: true
      createdAt:
        type: string
        format: date-time
        description: Document creation timestamp (ISO 8601)
      updatedAt:
        type: string
        format: date-time
        description: Document last update timestamp (ISO 8601)

  DeleteDocumentResponse:
    type: object
    required:
      - id
      - tenantId
      - status
    properties:
      id:
        type: string
        description: Document ID
      tenantId:
        type: string
        description: Tenant ID
      status:
        type: string
        enum: [deleted, not_found_or_deleted]
        description: Deletion status

  SearchResult:
    type: object
    required:
      - id
      - title
      - snippet
      - score
    properties:
      id:
        type: string
        description: Document ID
      title:
        type: string
        description: Document title
      snippet:
        type: string
        description: Highlighted snippet showing matching text fragments
      score:
        type: number
        format: float
        description: Relevance score (BM25)
      tags:
        type: array
        items:
          type: string
        description: Array of tags

  SearchResponse:
    type: object
    required:
      - tenantId
      - query
      - offset
      - limit
      - total
      - results
    properties:
      tenantId:
        type: string
        description: Tenant ID
      query:
        type: string
        description: Search query string used
      offset:
        type: integer
        minimum: 0
        description: Pagination offset
      limit:
        type: integer
        minimum: 1
        maximum: 50
        description: Number of results per page
      total:
        type: integer
        minimum: 0
        description: Total number of matching documents
      results:
        type: array
        items:
          $ref: '#/definitions/SearchResult'
        description: Array of search results

  Error:
    type: object
    required:
      - error
    properties:
      error:
        type: string
        description: Error message description